requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.0.2
google-api-core==1.25.1
google-auth==1.23.0
google-cloud-core==0.29.1
google-cloud-datastore==1.7.3
googleapis-common-protos==1.52.0
grpcio==1.35.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
WTForms==2.3.3
google-cloud-storage
flask_paginate==0.8.1
